Split .cpp files into classes, instead of general categories.
`Collider` now is similar to `Texture` in the sense that it has no constructor, and instead 4 creating functions; `Simple()`, `File()`, `Write()`, `ByTexture()`.
`ExpandMovementTree()` is no longer a member of `Object`.
Compressed the 16 vectors in `ExpandMovementTree()` into 4 of type `CollisionData`.
Added the variables required for simple mode in `Texture`, but functionality isn't yet implemented.
`Texture::File()` now returns a bool.