New texture file format standard.
The new texture file format doesn't use new lines ('\n') to distinguish the rows anymore. Instead, the first 8 bytes are 2 32 bits unsigned integers (X, then Y) representing the texture's fixed size, and the following data are continuous {RGBA foreground, RGBA background, char character} (9 bytes) structures representing the cells. `Texture::File()` now supports this, as well as the new function `Texture::ExportToFile()`.
`Engine::Print()` now prints '?' when a character is invalid (such us '\0' or '\n').
Unfinished Log() function, only in-engine function that uses it is `Texture::File()`. I will expand this in a later update.
Renamed `Input::Num()` to `Input::GetNum()`.
Repositioned Cell::c and CellA::c to be at the end (the 9th byte) which is important for texture importing and exporting.
`Texture::File()` now returns a UPoint of the new size of the texture.
Added `Texture::Resize()` which resizes the texture (simple or complex), and in the case of complex, doesn't overwrite the previous data in the new size, which is useful in some cases.
Added `Texture::ExportToFile()` which, as the name suggests, creates/opens a file and writes to it the texture using the texture file format standard.
Added some `rgb<ColorName>` macros in `usefulmacros.hpp` to make `Engine::Log()` a little easier to use.
Adapted widgets to use `Texture::SetForeground()` instead of doing it manually which saves a couple of lines of code and looks clearer.
Fixed `StringInputField` widget to use the given default value.