Reduced size of `Object`;
Compressed all of `Object`'s 9 "on event" callback pointers (std::function) to a single one, which receives an `Object::EventType` parameter stating the type of the event.
`Object` is now 160 bytes instead of 416. I considered also replacing the `Texture`s and `Collider`s vectors with arrays, but decided that the vectors are too comfortable to dispose of and `Object`s are now are small enough, I suppose.
Close #31.