Solving some GitHub issues;
Added `Objet::EnterLayer()` which makes the object remove itself from the last `parentLayer` and adds itself to the given layer. The "simpleplatform1" example tests this new functionality.
Close #23

Added `Time::CancelInvocation()` which takes a pointer returned by `Time::Invoke()`. The "widgetstest" example tests this new functionality.
Added `Input::RegisterRangedCallback()` which its handler is tremendously more optimized (speed and size) for handling a sequence of keys that are right next to each other on the ASCII table. `IntField` and `StringField` were updated to use this.
Close #22

Additionally:
	Fixed `Layer::RemoveObject()` functions, which didn't set the object's `parentLayer` to null, and removed the object which is one index before the correct object.
	Removed `instances` feature from `Time::Invocation`s as I have found it to be useless, and I always preferred invoking again at the end of the invoked function to create a loop.
	Changed `Time`'s invocations storage to be dynamic (pointers, `new`, `delete`).